Made ps1 compatible with mac, win

+++ tools/packup.ps1
@@ -1,5 +1,7 @@
 $rootDir = Split-Path $PSScriptRoot -Parent
-$distDir = Resolve-Path -Path "$rootDir\dist"
-$zipFile = Resolve-Path -Path "$rootDir\dist\Satr.zip"
-Remove-Item -Path $zipFile
+$distDir = Join-Path -Path "$rootDir" -ChildPath "dist"
+$zipFile = Join-Path -Path "$distDir" -ChildPath "Satr.zip"
+if (Test-Path $zipFile) {
+    Remove-Item -Path $zipFile
+}
 Compress-Archive -Path $distDir -DestinationPath $zipFile
